// WADNR.Web/src/app/shared/generated/model/project-update-organizations-step.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectOrganizationUpdateItem } from './project-organization-update-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectUpdateOrganizationsStep { 
    ProjectUpdateBatchID?: number;
    Organizations?: Array<ProjectOrganizationUpdateItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectUpdateOrganizationsStepForm { 
    ProjectUpdateBatchID?: FormControl<number>;
    Organizations?: FormControl<Array<ProjectOrganizationUpdateItem>>;
}

export class ProjectUpdateOrganizationsStepFormControls { 
    public static ProjectUpdateBatchID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Organizations = (value: FormControlState<Array<ProjectOrganizationUpdateItem>> | Array<ProjectOrganizationUpdateItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectOrganizationUpdateItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}