// WADNR.Web/src/app/shared/generated/model/project-update-organizations-step-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectOrganizationUpdateItemRequest } from './project-organization-update-item-request';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectUpdateOrganizationsStepRequest { 
    Organizations?: Array<ProjectOrganizationUpdateItemRequest> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectUpdateOrganizationsStepRequestForm { 
    Organizations?: FormControl<Array<ProjectOrganizationUpdateItemRequest>>;
}

export class ProjectUpdateOrganizationsStepRequestFormControls { 
    public static Organizations = (value: FormControlState<Array<ProjectOrganizationUpdateItemRequest>> | Array<ProjectOrganizationUpdateItemRequest> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectOrganizationUpdateItemRequest>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}