// WADNR.Web/src/app/shared/generated/model/project-organization-update-item-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectOrganizationUpdateItemRequest { 
    ProjectOrganizationUpdateID?: number | null;
    OrganizationID?: number;
    RelationshipTypeID?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectOrganizationUpdateItemRequestForm { 
    ProjectOrganizationUpdateID?: FormControl<number>;
    OrganizationID?: FormControl<number>;
    RelationshipTypeID?: FormControl<number>;
}

export class ProjectOrganizationUpdateItemRequestFormControls { 
    public static ProjectOrganizationUpdateID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static RelationshipTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}