// WADNR.Web/src/app/shared/generated/model/organization-lookup-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class OrganizationLookupItem { 
    OrganizationID?: number;
    OrganizationName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface OrganizationLookupItemForm { 
    OrganizationID?: FormControl<number>;
    OrganizationName?: FormControl<string>;
}

export class OrganizationLookupItemFormControls { 
    public static OrganizationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}